actualization general

+++ pages/gracias-por-comprar/index.tsx
@@ -1,9 +1,19 @@
+import { ISell } from '@/interfaces'
+import axios from 'axios'
 import React, { useEffect } from 'react'
 
 const PageBuySuccess = () => {
 
+  const updateClient = async () => {
+    if (localStorage.getItem('sell')) {
+      const sell: ISell = JSON.parse(localStorage.getItem('sell')!)
+      await axios.post('https://server-production-e234.up.railway.app/clients', { email: sell.email, firstName: sell.firstName, lastName: sell.lastName, phone: sell.phone, address: sell.address, departament: sell.details, region: sell.region, city: sell.city, tags: ['Clientes'] })
+      localStorage.clear()
+    }
+  }
+
   useEffect(() => {
-    localStorage.clear()
+    updateClient()
   }, [])
 
   return (
